Add files via upload
Lab 1: Array (Revision)
• Write a Java program that takes ten integers from the user and
constructs the following methods:
1. Write a method that displays the ten integers in reverse order of
their insertion.
2. Write a method that returns the max and min numbers.
3. Write a method that finds the minimum value in a range between
the 3rd number and the 8th number and prints it. Note, the 3rd
number and the 8th number are included in the range.
4. Write a method finding how many times a specific value is
repeated within the ten integers. If any such value is found,
display it and its number of repetitions, Example “7 is repeated 3
times”. If there are no duplicate elements, write the statement
“Your numbers are unique”.